+++ examples/example_6_bitwise_decoding.py
@@ -1,6 +1,6 @@
 """
-Epoch CCA and LDA
-==========================
+Bitwise Decoding
+================
 This script shows how to use CCA and LDA on epochs within a trial, using PyntBCI, for decoding c-VEP trials. Epochs are
 defined as the windows of data within a trial, synchronized to the onset of individual flashes. For each of these epochs
 the classifier determines whether a flash was presented or not. Integrating that information over time, allows the
@@ -95,16 +95,10 @@ plt.title("Single-trial labels")
 plt.tight_layout()
 
 # Visualize stimuli
-Vup = V.repeat(20, axis=1)  # upsample to better visualize the sharp edges
-plt.figure(figsize=(15, 8))
-plt.plot(np.arange(Vup.shape[1]) / (20 * fs), 2 * np.arange(n_classes) + Vup.T)
-for i in range(1 + int(V.shape[1] / (fs / fr))):
-    plt.axvline(i / fr, c="k", alpha=0.1)
-plt.yticks(2 * np.arange(n_classes), np.arange(n_classes))
-plt.xlabel("time [s]")
-plt.ylabel("code")
-plt.title("Code time-series")
-plt.tight_layout()
+fig, ax = plt.subplots(1, 1, figsize=(15, 8))
+pyntbci.plotting.stimplot(V, fs=fs, ax=ax, plotfs=False)
+fig.tight_layout()
+ax.set_title("Stimulus time-series")
 
 # %%
 # Epoch decoding